# Shrinker improvements