--- a/example/alcotest/output.txt.expected
+++ b/example/alcotest/output.txt.expected
@@ -9,30 +9,30 @@ Testing `my test'.
 ┌──────────────────────────────────────────────────────────────────────────────┐
 │ [FAIL]        suite              1   fail_sort_id.                           │
 └──────────────────────────────────────────────────────────────────────────────┘
-test `fail_sort_id` failed on ≥ 1 cases: [1; 0] (after 20 shrink steps)
-[exception] test `fail_sort_id` failed on ≥ 1 cases: [1; 0] (after 20 shrink steps)
+test `fail_sort_id` failed on ≥ 1 cases: [1; 0] (after 11 shrink steps)
+[exception] test `fail_sort_id` failed on ≥ 1 cases: [1; 0] (after 11 shrink steps)
  ──────────────────────────────────────────────────────────────────────────────
 ┌──────────────────────────────────────────────────────────────────────────────┐
 │ [FAIL]        suite              2   error_raise_exn.                        │
 └──────────────────────────────────────────────────────────────────────────────┘
 test `error_raise_exn`
 raised exception `Error`
-on `0 (after 63 shrink steps)`
+on `0 (after 1 shrink steps)`
 [exception] test `error_raise_exn`
 raised exception `Error`
-on `0 (after 63 shrink steps)`
+on `0 (after 1 shrink steps)`
  ──────────────────────────────────────────────────────────────────────────────
 ┌──────────────────────────────────────────────────────────────────────────────┐
 │ [FAIL]        suite              3   fail_check_err_message.                 │
 └──────────────────────────────────────────────────────────────────────────────┘
 test `fail_check_err_message` failed on ≥ 1 cases:
-0 (after 7 shrink steps)
+0 (after 1 shrink steps)
 this
 will
 always
 fail
 [exception] test `fail_check_err_message` failed on ≥ 1 cases:
-0 (after 7 shrink steps)
+0 (after 1 shrink steps)
 this
 will
 always
@@ -46,15 +46,9 @@ Test debug_shrink successfully shrunk counter example (step 0) to:
 (3, 1)
 ~~~ Shrink ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Test debug_shrink successfully shrunk counter example (step 1) to:
-(2, 1)
-~~~ Shrink ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Test debug_shrink successfully shrunk counter example (step 2) to:
-(2, 0)
-~~~ Shrink ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Test debug_shrink successfully shrunk counter example (step 3) to:
-(1, 0)
+(0, 1)
 law debug_shrink: 2 relevant cases (2 total)
-test `debug_shrink` failed on ≥ 1 cases: (1, 0) (after 3 shrink steps)
-[exception] test `debug_shrink` failed on ≥ 1 cases: (1, 0) (after 3 shrink steps)
+test `debug_shrink` failed on ≥ 1 cases: (0, 1) (after 1 shrink steps)
+[exception] test `debug_shrink` failed on ≥ 1 cases: (0, 1) (after 1 shrink steps)
  ──────────────────────────────────────────────────────────────────────────────
 4 failures! 6 tests run.