@jmid 2022-04-20
OK, using the merged shrinker performance benchmark I've now cherry-picked each of the four shrinker improvements and measured each of them.

* 5a67e37 - simpler list shrinker with better complexity
* d731af8 - use improved list shrinker for strings
* 43b3962 - improved function shrinker
* c19ccd0 - improved int shrinker

Executive summary:

- the improved `list`, `string`, and function shrinkers collectively reduce the runtime of the benchmark from 78.402s to 5.456s on my laptop
- adding the `int` shrinker improvement on top reduces the runtime from an average of 5.448s to 5.3289s across 10 runs

Whereas the latter represents a statistical significant improvement (measured with `ministat`) it is not as drastic as I thought before measuring.
The duplicate `0` testing of the improved `int` shrinker is unsatisfying, but so is the duplicate output of the current `int` shrinker:
  https://github.com/c-cube/qcheck/blob/824dafb111961ec16bc3526173ad004d45820632/test/core/QCheck_unit_tests.ml#L20-L24
We should address this separately. I therefore move to merge the first three commits.



Measurement details
-------------------

To keep the tables focused below I've deleted benchmark entries with a total QCheck1 shrinker time below 0.1 seconds.

Here's first the performance of current master. Note how the focused tables highlight string, list, and function benchmarks.
```
                                                         iteration seed 1234                   iteration seed 8743                   iteration seed 6789               total
Shrink test name                                  Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att    Q1/s   Q2/s
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
string never has a \255 char                      0.001  249/316     0.001   59/318     0.092 4466/4520    0.002   97/529     0.377 9260/9365    0.002   41/194     0.469  0.005
strings have unique chars                         0.003  248/269     0.000   18/30      0.919 4465/4536    0.002   24/52      0.000   14/34      0.000   15/20      0.922  0.002
lists shorter than 432                            6.343 1696/5118102  1.049  412/457     6.108 1612/4863421  0.981  405/450     6.087 1667/5037661  0.126  419/447    18.537  2.156
lists shorter than 4332                           2.172   13/190735  3.331 4022/4087    1.455   11/126052  3.580 4020/4067    1.410    7/126607  2.408 4013/4055    5.037  9.318
lists equal to duplication                        0.145   20/23      0.470    4/7       0.000    7/13      0.000    3/6       0.021   20/25      0.091   17/35      0.165  0.561
fold_left fold_right uncurried                    2.496   97/80630   0.044  376/1550    0.144   38/390     2.160 2064/8057    0.000    5/20      0.000    4/17      2.640  2.204
fold_left test, fun first                         0.001   40/57      0.001   15/28     40.082  191/44563   3.318   47/9773   10.359  223/75912   0.002   36/64     50.442  3.321
                                                                                                                                                                   78.402 19.754
```

**Step 1** Here's the performance of the simpler `list` shrinker. Note how the two `lists shorter than` benchmarks now run much faster.
On the other hand, the change causes `fold_left test, fun first` to really slow down because of the 93854 and 524691 shrink attempts with seed 8743 and 6789, respectively.
This latter test combines both lists, strings, and functions.

```
                                                         iteration seed 1234                   iteration seed 8743                   iteration seed 6789               total
Shrink test name                                  Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att    Q1/s   Q2/s
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
string never has a \255 char                      0.001  249/316     0.001   59/318     0.100 4466/4520    0.003   97/529     0.413 9260/9365    0.002   41/194     0.514  0.005
strings have unique chars                         0.003  248/269     0.000   18/30      0.917 4465/4536    0.001   24/52      0.000   14/34      0.000   15/20      0.920  0.002
lists shorter than 432                            0.161 1632/19461   0.903  412/457     0.194 1677/20004   1.016  405/450     0.076 1735/20729   0.112  419/447     0.431  2.032
lists shorter than 4332                           0.085   13/67      3.474 4022/4087    0.128   10/35      3.538 4020/4067    0.009    8/55      2.529 4013/4055    0.222  9.541
lists equal to duplication                        0.157   26/35      0.419    4/7       0.000    6/12      0.000    3/6       0.009   25/35      0.044   17/35      0.166  0.463
fold_left fold_right uncurried                    0.003   25/125     0.046  376/1550    0.023   34/119     2.255 2064/8057    0.000    5/20      0.000    4/17      0.025  2.301
fold_left test, fun first                         0.002   41/60      0.001   15/28     1488.644 1230/93854   3.423   47/9773   2123.311 2759/524691  0.002   36/64     3611.956  3.426
                                                                                                                                                                   3614.426 19.923
```

**Step 2** Here's the performance of using the improved `list` shrinker for `string`s. This reduces the runtime of the two string tests to just 0.001s. Note how for `strings have unique chars` with seed 8743 this reduces 4465 succesful shrink attempts out of 4536 down to just 15 out of 30.

The `string` improvement also helps a bit on `fold_left test, fun first` which is now down to 67879 and 124382 shrink attempts with seed 8743 and 6789, respectively.
```
                                                         iteration seed 1234                   iteration seed 8743                   iteration seed 6789               total
Shrink test name                                  Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att    Q1/s   Q2/s
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
string never has a \255 char                      0.000   14/28      0.001   59/318     0.001   15/24      0.005   97/529     0.001   21/38      0.003   41/194     0.001  0.008
strings have unique chars                         0.000   13/31      0.000   18/30      0.001   15/30      0.003   24/52      0.000    5/14      0.000   15/20      0.001  0.003
lists shorter than 432                            0.158 1632/19461   0.893  412/457     0.200 1677/20004   0.934  405/450     0.069 1735/20729   0.062  419/447     0.427  1.889
lists shorter than 4332                           0.140   13/67      3.365 4022/4087    0.116   10/35      3.214 4020/4067    0.018    8/55      2.391 4013/4055    0.273  8.970
lists equal to duplication                        0.160   26/35      0.382    4/7       0.000    6/12      0.000    3/6       0.009   25/35      0.103   17/35      0.169  0.485
fold_left fold_right uncurried                    0.003   25/125     0.046  376/1550    0.023   34/119     2.162 2064/8057    0.000    5/20      0.000    4/17      0.026  2.209
fold_left test, fun first                         0.017  275/543     0.001   15/28     669.151 1006/67879   3.552   47/9773   110.429  726/124382  0.002   36/64     779.596  3.555
                                                                                                                                                                   780.559 19.211
```

**Step 3** Here's the performance of adding the improved function shrinker on top.
This last one restores decent performance for `fold_left test, fun first` which is now down to 1551 and 1483 shrink attempts with seed 8743 and 6789, respectively.
```
                                                         iteration seed 1234                   iteration seed 8743                   iteration seed 6789               total
Shrink test name                                  Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att    Q1/s   Q2/s
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
string never has a \255 char                      0.000   14/28      0.001   59/318     0.000   15/24      0.005   97/529     0.001   21/38      0.003   41/194     0.002  0.008
strings have unique chars                         0.000   13/31      0.000   18/30      0.001   15/30      0.003   24/52      0.000    5/14      0.000   15/20      0.002  0.003
lists shorter than 432                            0.160 1632/19461   0.904  412/457     0.197 1677/20004   0.909  405/450     0.067 1735/20729   0.060  419/447     0.425  1.873
lists shorter than 4332                           0.135   13/67      3.534 4022/4087    0.116   10/35      3.196 4020/4067    0.020    8/55      2.364 4013/4055    0.270  9.094
lists equal to duplication                        0.156   26/35      0.381    4/7       0.000    6/12      0.000    3/6       0.009   25/35      0.103   17/35      0.165  0.483
fold_left fold_right uncurried                    0.002   44/199     0.044  376/1550    0.032   55/170     2.106 2064/8057    0.000    5/15      0.000    4/17      0.034  2.151
fold_left test, fun first                         0.016  275/543     0.000   15/28      4.206  305/1551    3.254   47/9773    0.272  296/1483    0.002   36/64      4.494  3.257
                                                                                                                                                                    5.456 18.971
```

**Step 4** Here's the performance of adding the improved `int` shrinker on top.
Note how `lists shorter than 432` has its total shrink attempt reduced from 19461, 20004, and 20729 down to 4881, 4728, 4793 total attempts.
```
                                                         iteration seed 1234                   iteration seed 8743                   iteration seed 6789               total
Shrink test name                                  Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att   Q1/s  #succ/#att   Q2/s  #succ/#att    Q1/s   Q2/s
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
string never has a \255 char                      0.000   14/28      0.001   59/318     0.000   15/24      0.003   97/529     0.001   21/38      0.002   41/194     0.002  0.005
strings have unique chars                         0.000   13/31      0.000   18/30      0.001   15/30      0.002   24/52      0.000    5/14      0.000   15/20      0.001  0.002
lists shorter than 432                            0.107  417/4881    0.831  412/457     0.143  404/4728    0.871  405/450     0.021  407/4793    0.060  419/447     0.271  1.762
lists shorter than 4332                           0.133   13/67      3.335 4022/4087    0.083   10/35      3.583 4020/4067    0.010    8/55      2.265 4013/4055    0.226  9.183
lists equal to duplication                        0.166   20/23      0.385    4/7       0.000    3/6       0.000    3/6       0.010   18/21      0.101   17/35      0.176  0.487
fold_left fold_right uncurried                    0.002   39/233     0.043  376/1550    0.026   49/161     2.119 2064/8057    0.000    2/14      0.000    4/17      0.028  2.163
fold_left test, fun first                         0.017  272/534     0.001   15/28      4.225  305/1558    3.208   47/9773    0.274  288/1434    0.002   36/64      4.516  3.211
                                                                                                                                                                    5.241 18.957
```



**Statistical test** For reference here's the total timings of 10 runs of step 3:
```
5.456
5.508
5.441
5.442
5.406
5.414
5.417
5.454
5.435
5.507
```
and 10 runs of step 4:
```
5.241
5.232
5.269
5.240
5.376
5.289
5.405
5.413
5.416
5.408
```
and the output of `ministat`:
```
x shrink_bench.log-stepwise-step3-improved-function-shrinker-timings
+ shrink_bench.log-stepwise-step4-improved-int-shrinker-timings
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|+    ++                 +            +                                                       +                  * +  * *           x   x       xx                                xx|
|          |___________________________________________________A______________________________M_____________________||__________________M___A______________________|                |
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
    N           Min           Max        Median           Avg        Stddev
x  10         5.406         5.508         5.442         5.448   0.035458896
+  10         5.232         5.416         5.376        5.3289   0.081052041
Difference at 95.0% confidence
	-0.1191 +/- 0.0587783
	-2.18612% +/- 1.0789%
	(Student's t, pooled s = 0.062557)
```